增加分配权限接口

+++ springcloud-mvc/src/main/java/com/springcloud/mvc/controller/role/RoleController.java
@@ -1,9 +1,11 @@
 package com.springcloud.mvc.controller.role;
 
 import com.role.api.RoleFeignApi;
+import com.role.api.UserRoleFeignApi;
 import common.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -21,6 +23,8 @@ public class RoleController {
 
     @Resource
     private RoleFeignApi roleFeignApi;
+    @Resource
+    private UserRoleFeignApi userRoleFeignApi;
 
     /**
      * 删除一个角色
@@ -28,8 +32,15 @@ public class RoleController {
      * @param roleId 角色id
      * @return Result<Void>
      */
+    @Transactional(rollbackFor = Exception.class)
     @DeleteMapping(value = "/mvc/role")
-    Result<Void> deleteRole(@RequestParam(value = "roleId") Integer roleId) {
+    Result<Void> deleteRole(@RequestParam(value = "roleId", required = false) Integer roleId,
+                            @RequestParam(value = "uid", required = false) Integer uid) {
+        //删除用户角色关联关系
+        Result<Void> userRoleResult = userRoleFeignApi.deleteUserRole(roleId, uid);
+        if (null == roleId) {
+            return userRoleResult;
+        }
         return roleFeignApi.deleteRole(roleId);
     }
 }
